refactor: 삭제 버튼 컴포넌트 추가
